pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "monocle3-python"
version = "1.4.26+4f4239a"
description = "Python port of the R monocle3 package (AnnData-based)."
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "pandas>=2.0",
    "anndata>=0.10",
    "scanpy>=1.10",
    "scikit-learn>=1.3",
    "umap-learn>=0.5",
    "openTSNE>=1.0",
    "python-igraph>=0.10",
    "leidenalg>=0.10",
    "pynndescent>=0.5",
    "hnswlib>=0.8",
    "scanorama>=1.7",
    "statsmodels>=0.14",
    "glum>=3.0",
    "libpysal>=4.10",
    "esda>=2.5",
    "patsy>=0.5",
    "plotly>=5.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "ruff"]
docs = ["mkdocs", "mkdocstrings[python]", "mkdocs-material", "mkdocs-jupyter"]

[tool.hatch.build.targets.wheel]
packages = ["monocle3"]

[tool.hatch.build.targets.sdist]
include = [
    "/monocle3",
    "/tests",
    "/tutorials",
    "/validation",
    "/docs",
    "/README.md",
    "/LICENSE",
    "/mkdocs.yml",
    "/pyproject.toml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
